Remove asset catalogue properties
