+++ server/graphql/asset_catalogue/src/lib.rs
@@ -1,6 +1,3 @@
-use asset_catalogue_property_queries::{
-    asset_catalogue_properties, AssetCataloguePropertyFilterInput,
-};
 use async_graphql::*;
 use graphql_core::pagination::PaginationInput;
 use mutations::{
@@ -9,14 +6,12 @@ use mutations::{
 };
 use types::{
     asset_catalogue_item::{AssetCatalogueItemResponse, AssetCatalogueItemsResponse},
-    asset_catalogue_property::AssetCataloguePropertyResponse,
     asset_category::{AssetCategoriesResponse, AssetCategoryResponse},
     asset_class::{AssetClassResponse, AssetClassesResponse},
     asset_type::{AssetTypeResponse, AssetTypesResponse},
 };
 
 pub mod asset_catalogue_item_queries;
-mod asset_catalogue_property_queries;
 use crate::asset_catalogue_item_queries::*;
 pub mod asset_category_queries;
 use crate::asset_category_queries::*;
@@ -101,14 +96,6 @@ impl AssetCatalogueQueries {
 
     pub async fn asset_type(&self, ctx: &Context<'_>, id: String) -> Result<AssetTypeResponse> {
         asset_type(ctx, id)
-    }
-
-    pub async fn asset_catalogue_properties(
-        &self,
-        ctx: &Context<'_>,
-        #[graphql(desc = "Filter option")] filter: Option<AssetCataloguePropertyFilterInput>,
-    ) -> Result<AssetCataloguePropertyResponse> {
-        asset_catalogue_properties(ctx, filter)
     }
 }
