+++ server/graphql/asset_catalogue/src/types/mod.rs
@@ -1,5 +1,4 @@
 pub mod asset_catalogue_item;
-pub mod asset_catalogue_property;
 pub mod asset_category;
 pub mod asset_class;
 pub mod asset_type;
