+++ server/repository/src/db_diesel/assets/mod.rs
@@ -1,8 +1,6 @@
 pub mod asset;
 pub mod asset_catalogue_item;
 pub mod asset_catalogue_item_row;
-pub mod asset_catalogue_property;
-pub mod asset_catalogue_property_row;
 pub mod asset_category;
 pub mod asset_category_row;
 pub mod asset_class;
