+++ server/service/src/sync/test/test_data/mod.rs
@@ -5,7 +5,6 @@ use super::{TestSyncIncomingRecord, TestSyncOutgoingRecord};
 pub(crate) mod activity_log;
 pub(crate) mod asset;
 pub(crate) mod asset_catalogue_item;
-pub(crate) mod asset_catalogue_property;
 pub(crate) mod asset_category;
 pub(crate) mod asset_class;
 pub(crate) mod asset_log;
@@ -75,7 +74,6 @@ pub(crate) fn get_all_pull_upsert_central_test_records() -> Vec<TestSyncIncoming
     test_records.append(&mut asset_category::test_pull_upsert_records());
     test_records.append(&mut asset_type::test_pull_upsert_records());
     test_records.append(&mut asset_catalogue_item::test_pull_upsert_records());
-    test_records.append(&mut asset_catalogue_property::test_pull_upsert_records());
     test_records.append(&mut asset::test_pull_upsert_records());
     test_records.append(&mut asset_log::test_pull_upsert_records());
     test_records.append(&mut asset_log_reason::test_pull_upsert_records());
@@ -163,7 +161,6 @@ pub(crate) fn get_all_sync_v6_records() -> Vec<TestSyncOutgoingRecord> {
     test_records.append(&mut asset_category::test_v6_central_push_records());
     test_records.append(&mut asset_type::test_v6_central_push_records());
     test_records.append(&mut asset_catalogue_item::test_v6_central_push_records());
-    test_records.append(&mut asset_catalogue_property::test_v6_central_push_records());
     test_records.append(&mut asset::test_v6_records());
     test_records.append(&mut asset_log::test_v6_records());
     test_records.append(&mut asset_log_reason::test_v6_records());
